requirements.txt
python-dotenv
delegator.py
datetime
cryptography